Add: stm32mp1 publish script

+++ publish/seeed-stm32mp1-stable.sh
@@ -3,24 +3,31 @@
 time=$(date +%Y-%m-%d)
 DIR="$PWD"
 
-ssh_user="buildbot@beagleboard.org"
+ssh_svr=192.168.13.13
+ssh_user="seeeder@${ssh_svr}"
 rev=$(git rev-parse HEAD)
 branch=$(git describe --contains --all HEAD)
-server_dir="/var/lib/buildbot/masters/kernel-buildbot/public_html/images/${branch}/${rev}"
+server_dir="/home/public/share/stm32mp1"
+this_name=$0
 
 export apt_proxy=localhost:3142/
 
 keep_net_alive () {
 	while : ; do
 		sleep 15
-		echo "log: [Running: ./publish/bb.org_4gb_stable.sh]"
+		echo "log: [Running: ${this_name}]"
 	done
 }
 
 build_and_upload_image () {
 	echo "***BUILDING***: ${config_name}: ${target_name}-${image_name}-${size}.img"
-	
-	./RootStock-NG.sh -c ${config_name}
+
+	# To prevent rebuilding:
+	# export FULL_REBUILD=
+	FULL_REBUILD=${FULL_REBUILD-1}
+	if [ -n "${FULL_REBUILD}" -o ! -e "deploy/${image_name}.tar" ]; then
+		./RootStock-NG.sh -c ${config_name}
+	fi
 
 	if [ -d ./deploy/${image_name} ] ; then
 		cd ./deploy/${image_name}/
@@ -28,8 +35,11 @@ build_and_upload_image () {
 		sudo ./setup_sdcard.sh ${options}
 
 		if [ -f ${target_name}-${image_name}-${size}.img ] ; then
-			sudo chown buildbot.buildbot ${target_name}-${image_name}-${size}.img
-			sudo chown buildbot.buildbot ${target_name}-${image_name}-${size}.img.xz.job.txt
+			me=`whoami`
+			sudo chown ${me}.${me} ${target_name}-${image_name}-${size}.img
+			if [ -f "${target_name}-${image_name}-${size}.img.xz.job.txt" ]; then
+				sudo chown ${me}.${me} ${target_name}-${image_name}-${size}.img.xz.job.txt
+			fi
 
 			sync ; sync ; sleep 5
 
@@ -42,12 +52,16 @@ build_and_upload_image () {
 			ssh ${ssh_user} mkdir -p ${server_dir}
 			rsync -e ssh -av ./${target_name}-${image_name}-${size}.bmap ${ssh_user}:${server_dir}/
 			rsync -e ssh -av ./${target_name}-${image_name}-${size}.img.xz ${ssh_user}:${server_dir}/
-			rsync -e ssh -av ./${target_name}-${image_name}-${size}.img.xz.job.txt ${ssh_user}:${server_dir}/
+			if [ -f "${target_name}-${image_name}-${size}.img.xz.job.txt" ]; then
+				rsync -e ssh -av ./${target_name}-${image_name}-${size}.img.xz.job.txt ${ssh_user}:${server_dir}/
+			fi
 			rsync -e ssh -av ./${target_name}-${image_name}-${size}.img.xz.sha256sum ${ssh_user}:${server_dir}/
 
 			#cleanup:
 			cd ../../
-			sudo rm -rf ./deploy/ || true
+
+			# TODO
+			# sudo rm -rf ./deploy/ || true
 		else
 			echo "***ERROR***: Could not find ${target_name}-${image_name}-${size}.img"
 		fi
@@ -59,47 +73,20 @@ build_and_upload_image () {
 keep_net_alive & KEEP_NET_ALIVE_PID=$!
 echo "pid: [${KEEP_NET_ALIVE_PID}]"
 
-# IoT BeagleBone image
-##Debian 8:
+# Console STM32MP1 image
+##Debian 10:
 #image_name="${deb_distribution}-${release}-${image_type}-${deb_arch}-${time}"
-image_name="debian-8.11-iot-2gb-armhf-${time}"
+image_name="debian-buster-console-armhf-${time}"
 size="2gb"
-target_name="bone"
-options="--img-2gb ${target_name}-${image_name} --dtb beaglebone \
---hostname beaglebone --enable-cape-universal"
-config_name="bb.org-debian-jessie-iot-2gb-v4.4"
+target_name="stm32mp1"
+options="--img-2gb ${target_name}-${image_name} --hostname stm32mp1 --dtb stm32mp1"
+config_name="seeed-debian-buster-console-v4.19"
 build_and_upload_image
 
-# LXQT BeagleBone image
-##Debian 8:
-#image_name="${deb_distribution}-${release}-${image_type}-${deb_arch}-${time}"
-image_name="debian-8.11-lxqt-4gb-armhf-${time}"
-size="4gb"
-target_name="bone"
-options="--img-4gb ${target_name}-${image_name} --dtb beaglebone \
---hostname beaglebone --enable-cape-universal"
-config_name="bb.org-debian-jessie-lxqt-4gb-v4.4"
-build_and_upload_image
+[ -e /proc/$KEEP_NET_ALIVE_PID ] && {
+	# TODO
+	# sudo rm -rf ./deploy/ || true
+	sudo kill $KEEP_NET_ALIVE_PID
+}
 
-# LXQT BeagleBoard-xM image
-##Debian 8:
-#image_name="${deb_distribution}-${release}-${image_type}-${deb_arch}-${time}"
-image_name="debian-8.11-lxqt-xm-4gb-armhf-${time}"
-size="4gb"
-target_name="bbxm"
-options="--img-4gb ${target_name}-${image_name} --dtb omap3-beagle-xm --rootfs_label rootfs --hostname beagleboard"
-config_name="bb.org-debian-jessie-lxqt-4gb-xm"
-build_and_upload_image
-
-# LXQT BeagleBoard-X15 image
-##Debian 8:
-#image_name="${deb_distribution}-${release}-${image_type}-${deb_arch}-${time}"
-image_name="debian-8.11-lxqt-4gb-armhf-${time}"
-size="4gb"
-target_name="am57xx"
-options="--img-4gb ${target_name}-${image_name} --dtb am57xx-beagle-x15 --hostname beagleboard"
-config_name="bb.org-debian-jessie-lxqt-4gb-v4.4"
-build_and_upload_image
-
-[ -e /proc/$KEEP_NET_ALIVE_PID ] && sudo kill $KEEP_NET_ALIVE_PID
